Break down gui into modules
